Implement stricter type checking when registering subscribers
