+++ api/src/main/java/net/kyori/event/SimpleEventBus.java
@@ -23,6 +23,7 @@
  */
 package net.kyori.event;
 
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.SetMultimap;
 import org.checkerframework.checker.nullness.qual.NonNull;
@@ -33,10 +34,16 @@ import java.util.function.Predicate;
  * A simple implementation of an event bus.
  */
 public class SimpleEventBus<E> implements EventBus<E> {
+  private final Class<E> type;
   private final SubscriberRegistry<E> registry = new SubscriberRegistry<>();
+
+  public SimpleEventBus(final @NonNull Class<E> type) {
+    this.type = type;
+  }
 
   @Override
   public <T extends E> void register(final @NonNull Class<T> clazz, final @NonNull EventSubscriber<? super T> subscriber) {
+    Preconditions.checkArgument(this.type.isAssignableFrom(clazz), "clazz " + clazz + " cannot be casted to event type " + this.type);
     this.registry.register(clazz, subscriber);
   }
 
@@ -58,6 +65,12 @@ public class SimpleEventBus<E> implements EventBus<E> {
   @Override
   public @NonNull SetMultimap<Class<?>, EventSubscriber<?>> subscribers() {
     return this.registry.subscribers();
+  }
+
+  @Override
+  public <T extends E> boolean hasSubscribers(final @NonNull Class<T> clazz) {
+    Preconditions.checkArgument(this.type.isAssignableFrom(clazz), "clazz " + clazz + " cannot be casted to event type " + this.type);
+    return !this.registry.subscribers(clazz).isEmpty();
   }
 
   @Override
@@ -88,7 +101,7 @@ public class SimpleEventBus<E> implements EventBus<E> {
   }
 
   @Override
-  public <T extends E> boolean hasSubscribers(final @NonNull Class<T> clazz) {
-    return !this.registry.subscribers(clazz).isEmpty();
+  public @NonNull Class<E> eventType() {
+    return this.type;
   }
 }
