+++ method/build.gradle
@@ -1,6 +1,5 @@
 dependencies {
   compile project(':event-api')
-  compile 'org.slf4j:slf4j-api:1.7.25'
 }
 
 jar {
